requirements.txt
git+https://github.com/Wyattjoh/pushover
pyyaml
selenium
